Invasive rewrite to defaulting rules.
There are 2 main parts to this:

1. Element-scoped defaults:

Previously, I was attempting to implement element-scoped defaults atop the
existing public CF APIs. You can see some of that here:
https://github.com/jspecify/nullness-checker-for-checker-framework/blob/fda816f563f7cd8f79dea6e4ec0b3c171befff1a/src/main/java/com/google/jspecify/nullness/NullSpecAnnotatedTypeFactory.java#L611

This turns out not to work in some cases. Notably, CF does not always run
`annotate(foo)` before `annotate(elementEnclosedByFoo)`. (I had wondered about
this, and I am surprised that I got away with depending on it for so long!)

I gave a little thought to how to expose my new feature most elegantly while
preserving existing functionality. But for the moment, I decided to just
replace existing functionality, including:

- I removed support for `@DefaultQualifier`. Removing support is probably good
from the perspective of cross-tool compatibility, but it's a regression for CF
users.

- When inheriting a `DefaultSet` from a parent element, I *always* copy it.
This is the safest easy way to let sub-elements manipulate the set
programmatically, but it's wasteful.

- To support the former bullet, I now permit new defaults to override existing
defaults. My forthcoming JSpecify changes will require this in order to avoid
an exception when calling the public `addElementDefault` API. Note, though,
that I suspect there are bugs that arise in normal CF usage but _don't_ produce
exceptions because they operate directly on the `DefaultSet`. See Checker
Framework issue number 3677.

- I added `protected void populateNewDefaults(Element elt, boolean
initialDefaultsAreEmpty)`.

- I no longer inherit defaults from parent packages.

2. Type-variable defaulting

I added `protected boolean
shouldAnnotateOtherwiseNonDefaultableTypeVariable(AnnotationMirror qual,
boolean isDeclaration)`. Its usage is a little subtle. In particular, if the
per-element defaults decline to annotate the type variable, the overall
checked-code defaults (which run later) will then get another shot. This may or
may not be what's desired. In my case, it's not desired. My workaround is going
to be to avoid setting overall checked-code defaults in favor of setting
defaults on each top-level class (hence the `boolean initialDefaultsAreEmpty`
parameter to the `populateNewDefaults` method discussed above).