Experiment with omitting type-variable bounds from error messages.
Unlike stock CF, we do not modify bounds when applying an annotation to
a type-variable usage. Accordingly, we print such an annotation on the
type-variable usage itself (5fc6c2807617eeaf415d2acd69fc11ace50d98d5).

Thus, there is somewhat less need to print bounds, too. And omitting
them reduces the length of some messages significantly. That said,
knowing the bounds is extremely useful in some cases. The right
compromise may be to print them as a sort of "footnote," similar to what
javac does for type variables or wildcards or both(??).

But for now, let's see how far we can get with the simple thing.