Further changes in support of defaulting of unbounded wildcards.
I ran into problems with references to unbounded wildcards _written in
bytecode_, such as Guava's usage of the currently unannotated
java.lang.reflect.Array:

https://github.com/google/guava/blob/751d7c0d5f1fe3cbc067987e7d185763affa9ec7/guava/src/com/google/common/collect/ObjectArrays.java#L49

In those cases, wildcard.bound apparently had been populated (by CF?).
Then, presumably because of my earlier change to CF, CF was concluding
that, given that there is a bound, then if the extends bound was
implicit, then the super bound must be *explicit*? Or something like
that. I don't think I've ever fully understood the logic I was messing
with, and it seems to have bit me here.

Anyway, it presumably makes sense to trust wildcard.isUnbound() in
general, given the simpler(?) approach to wildcards I've put into CF.
And trusting it fixes the problem case while not breaking anything else
in Guava or in the sample inputs.