Don't inherit annotations from supermethods and supertypes.
It's not clear what behavior we'll want in the long term:
https://github.com/jspecify/jspecify/issues/155

In the short term, this buys us 2 things:

1. It lets us check that we're not currently depending on that behavior
in our samples (or in Guava, which I'm also testing against).

2. It prevents an error from arising when checking Guava after
https://github.com/jspecify/nullness-checker-for-checker-framework/commit/2bab27eb167dcfb5ac970d837c0941151282939d:

"GenericAnnotatedTypeFactory.addComputedTypeAnnotations: root needs to
be set when used on trees"

The error is probably my fault: I am calling
`isNullExclusiveUnderEveryParameterization` during
`substituteTypeVariable`. That triggers `populateNewDefaults`, which
eventually calls into `inheritOverriddenDeclAnnos`, which eventually
calls into `directSuperTypes`, which eventually calls into
`addComputedTypeAnnotations`.

What I am doing may be fundamentally bogus. But for now, I can avoid it
by eliminating the `inheritOverriddenDeclAnnos` part of the call. And
once I'm doing that, I might as well eliminate inheritance entirely so
that we're consistent about it.

And in fact, not only does this eliminate that exception, but it also
eliminates the infinite recursion that I referred to in my description
for the commit linked above.