Reliably read annotations from stub files.
CORRECTION: Contrary to what I said in my previous commit's message, I
_do_ still need to include changes to prevent infinite recursion. (I had
gotten confused when testing. Sorry about that.) The earlier commit is
still useful for the other reasons discussed in its message, just not
for this specific reason. In particular, it is still necessary in order
to prevent this commit from triggering the "root needs to be set" error.

(That said, I am still confused by at least one result of this commit.
So maybe I continue to have it wrong.)

As for this commit: I believe that my earlier description of the problem
it's solving was correct:
https://github.com/jspecify/nullness-checker-for-checker-framework/commit/2bab27eb167dcfb5ac970d837c0941151282939d

This commit implements the second possible solution I described
("Another possible way forward...").