Include the class name in errors about method arguments.
The particular case we got a report about was about the worst possible
case:

```
incompatible argument for parameter value of of.
```

This was a report about Optional.of, whose parameter is named "value."

Now we would see:

```
incompatible argument for parameter value of Optional.of.
```

A complicating factor is that the caret in the error message pointed at
the parenthesis of the call that *produced* the argument, not at
the call to Optional.of itself or the beginning of that argument. I'm
not sure what the best approach to the caret would be in general, but
hopefully this tweak to the message helps mitigate the current caret
placement.