Fix constructor return types' type arguments.
Since we apply defaults to type-variable usages, we're also applying
them to the `T` in `Foo<T>`, the return type of constructor `Foo()`.

That doesn't actually break handling of any sample inputs, but it was
confusing me while I was debugging something else.

(Of course, it's always possible that, by fixing this bug, I've made it
possible to remove some existing workaround that I didn't fully
understand when I put it in.)