Fix defaulting for declarations' type arguments, too.
This is similar to the previous commit.

And, like the previous commit, it doesn't change behavior for any
existing samples.

Unlike the previous commit, it at least changes some behavior in
combination with some other changes I've been experimenting with. I
still don't know that those changes will ever see the light of day,
though. So a large part of my motivation is just to get correct types in
more places in case it matters "sometime."

This commit also tweaks the signature of
shouldAnnotateOtherwiseNonDefaultableTypeVariable. I will update the
checker accordingly.