Ignore annotations on type parameters themselves.
(just as 1db27bac27507a46a9cc7a894df5e9c8d64c2722 did for wildcards)

I should have noted some attempted alternatives on that prior commit:

- I tried to do this in TreeAnnotator, but TreeAnnotator "does not visit
component parts of the tree." That appears to mean that it does not even
call visitDeclaredType, let alone visitWildcard. Plus:

- I tried to do this in TypeAnnotator, but possibly it was happening too
late? Or something? (The same problem might affect TreeAnnotator.) I
didn't dig deeply. Ah, now that I'm writing this, I suspect that the
problem may be that I tried to remove the annotation from the wildcard
type but not from the bound type that CF actually attaches it to? So
maybe I could do this in our checker, rather than in CF itself. For the
moment, I'll stick with this.